add new inventory store

+++ omod/src/main/java/org/openmrs/module/ehrinventory/web/controller/store/StoreFormController.java
@@ -3,12 +3,12 @@
  * Version 1.0 (the "License"); you may not use this file except in
  * compliance with the License. You may obtain a copy of the License at
  * http://license.openmrs.org
- *
+ * <p>
  * Software distributed under the License is distributed on an "AS IS"
  * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
  * License for the specific language governing rights and limitations
  * under the License.
- *
+ * <p>
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 package org.openmrs.module.ehrinventory.web.controller.store;
@@ -48,8 +48,8 @@ public class StoreFormController {
 
 	@RequestMapping(method = RequestMethod.GET)
 	public String firstView(@ModelAttribute("store") InventoryStore store,
-			@RequestParam(value = "storeId", required = false) Integer id,
-			Model model) {
+							@RequestParam(value = "storeId", required = false) Integer id,
+							Model model) {
 		if (id != null) {
 			store = Context.getService(InventoryService.class).getStoreById(id);
 			model.addAttribute("store", store);
@@ -59,51 +59,48 @@ public class StoreFormController {
 		model.addAttribute("listIsDrug", listIsDrug);
 		return "/module/ehrinventory/store/form";
 	}
-	@ModelAttribute("roles")
-	public List<Role> populateRoles(HttpServletRequest request,Model model) {
 
-		 
-	
-		// return listRole;
-		
+	@ModelAttribute("roles")
+	public List<Role> populateRoles(HttpServletRequest request, Model model) {
+
+		List<Role> roles = Context.getUserService().getAllRoles();
+		List<Role> listRole = new ArrayList<Role>();
+		listRole.addAll(roles);
+
 		InventoryService inventoryService = (InventoryService) Context.getService(InventoryService.class);
+		List<InventoryStore> stores = inventoryService.listAllInventoryStore();
+
+		List<InventoryStoreRoleRelation> relation = inventoryService.listInventoryStoreRole();
+
+
 		String storeId = request.getParameter("storeId");
 		String role = "";
 		InventoryStore store = null;
 		InventoryStoreRoleRelation srl = null;
-		if( storeId != null )
-		{
+		if (storeId != null) {
+			//List<InventoryStoreRoleRelation> storeRelation = inventoryService.listOfRoleRelationStore(Integer.parseInt(storeId));
+			List<InventoryStoreRoleRelation> storeRelation = inventoryService.listOfRoleRelationStore(NumberUtils.toInt(storeId));
+			model.addAttribute("selectedModule", storeRelation);
 			store = inventoryService.getStoreById(NumberUtils.toInt(storeId));
-			if(store != null){
+			if (store != null) {
 				role = store.getRole() != null ? store.getRole().getRole() : "";
 			}
-			
-		}
-		
-		//List<InventoryStoreRoleRelation> storeRelation = inventoryService.listOfRoleRelationStore(Integer.parseInt(storeId));
-		List<InventoryStoreRoleRelation> storeRelation = inventoryService.listOfRoleRelationStore(NumberUtils.toInt(storeId));
-		model.addAttribute("selectedModule", storeRelation);
+			if (!CollectionUtils.isEmpty(roles) && !CollectionUtils.isEmpty(stores)) {
+				for (Role roleX : roles) {
+					for (InventoryStoreRoleRelation sr : relation) {
+						if (roleX.getRole().equals(sr.getRoleName()) && !(sr.getStoreid().equals(store.getId()))) {
 
-		List<Role> roles = Context.getUserService().getAllRoles();
-		List<InventoryStoreRoleRelation> relation = inventoryService.listInventoryStoreRole();
-		
-		List<Role> listRole = new ArrayList<Role>();
-		listRole.addAll(roles);
-		
-	    List<InventoryStore> stores = inventoryService.listAllInventoryStore();
-	  
-	    if(!CollectionUtils.isEmpty(roles) && !CollectionUtils.isEmpty(stores)){
-		    for( Role roleX : roles ){
-		    	for( InventoryStoreRoleRelation sr : relation )	{
-		    		if( roleX.getRole().equals(sr.getRoleName()) && !(sr.getStoreid().equals(store.getId()))){
-		    			
-		    			listRole.remove(roleX);
-		    		}
-		    	}
-		    }
-	    }
-	    return listRole;
-		
+							listRole.remove(roleX);
+						}
+					}
+				}
+			}
+
+		}
+
+
+		return listRole;
+
 	}
 
 	@ModelAttribute("parents")
@@ -113,10 +110,10 @@ public class StoreFormController {
 		List<InventoryStore> stores = inventoryService.listAllInventoryStore();
 		List<InventoryStore> listParents = new ArrayList<InventoryStore>();
 		listParents.addAll(stores);
-		
-		for (InventoryStore storeV : stores) { 
+
+		for (InventoryStore storeV : stores) {
 			if (storeV.getParentStores().size() > 0) { // luan. should be
-				 								// updated here
+				// updated here
 				listParents.remove(storeV);
 			}
 
@@ -136,8 +133,8 @@ public class StoreFormController {
 	 */
 	@RequestMapping(method = RequestMethod.POST)
 	public String onSubmit(@ModelAttribute("store") InventoryStore store,
-			
-			Model model, HttpServletRequest request, SessionStatus status) {
+
+						   Model model, HttpServletRequest request, SessionStatus status) {
 		// new StoreValidator().validate(store, bindingResult);
 		// storeValidator.validate(store, bindingResult);
 
@@ -152,7 +149,7 @@ public class StoreFormController {
 				&& !request.getParameterValues("parent")[0].isEmpty()) {
 			Set<InventoryStore> listParents = new HashSet<InventoryStore>();
 			for (String parentId : request.getParameterValues("parent")) {
-				
+
 
 				InventoryStore parentStore = new InventoryStore();
 				parentStore = inventoryService.getStoreById(Integer
@@ -163,36 +160,28 @@ public class StoreFormController {
 			store.setParentStores(listParents);
 
 		}
-		
 		//Multiple role to store
 		InventoryStoreRoleRelation rel = new InventoryStoreRoleRelation();
-		int value=0;
 		String roleName = request.getParameter("roles");
-		
-		List<InventoryStoreRoleRelation> roleRelation = inventoryService.listOfRoleRelation(store.getId(),roleName);
-		
-		for (InventoryStoreRoleRelation rl : roleRelation) {
-			inventoryService.deleteStoreRole(rl);
-		}
 
-			if(value==0)
-				
-				{ 	
-				for( String roleName1 :request.getParameterValues("roles"))
-					
-				{ 
+		if (store.getId() != null) {
+			List<InventoryStoreRoleRelation> roleRelation = inventoryService.listOfRoleRelation(store.getId(), roleName);
+			for (InventoryStoreRoleRelation rl : roleRelation) {
+				inventoryService.deleteStoreRole(rl);
+			}
+
+
+			for (String roleName1 : request.getParameterValues("roles")) {
 				rel.setRoleName(roleName1);
 				rel.setStoreid(store.getId());
 				inventoryService.saveStores(rel);
-				
-				}
-			
-				}
+			}
 
-		
-
+		}
 		inventoryService.saveStore(store);
-		 status.setComplete();
+
+
+		status.setComplete();
 		return "redirect:/module/ehrinventory/storeList.form";
 	}
 }
